Undo remesh from v1.4, shade flat, fill holes
The remesh from v1.4 was unecessary because the
slicer bugs from v1.3 were due to holes in the
shoulder of the right arm causing the slicer
import to appear without the face that the peg
sits on. The awkward shadows/textures mentioned
in v1.4 on the torso and arms were also due
to smooth shading so in this version all pegs and
their faces were shaded flat on the torso and
arms. Only holes from the right arm were filled
in this version.